refactor: refactor DrawBoard components
